move modern LargeFiatTokenValueView
