+++ src/components/display/LargeFiatTokenValueView.js
